+++ Libplanet.Tests/HashcashTest.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Linq;
 using System.Security.Cryptography;
 using Libplanet.Blocks;
@@ -15,9 +16,18 @@ namespace Libplanet.Tests
         public void AnswerSatisfiesDifficulty(byte[] challenge, long difficulty)
         {
             byte[] Stamp(Nonce nonce) => challenge.Concat(nonce.ToByteArray()).ToArray();
-            var answer = Hashcash.Answer(Stamp, difficulty);
-            var digest = Hashcash.Hash(Stamp(answer));
-            Assert.True(digest.Satisfies(difficulty));
+            (Nonce answer, ImmutableArray<byte> digest) =
+                Hashcash.Answer(Stamp, HashAlgorithmType.Of<SHA256>(), difficulty);
+            Assert.True(Satisfies(digest.ToArray(), difficulty));
+            TestUtils.AssertBytesEqual(
+                digest.ToArray(),
+                SHA256.Create().ComputeHash(Stamp(answer))
+            );
+#pragma warning disable CS0612
+            Nonce answer2 = Hashcash.Answer(Stamp, difficulty);
+            BlockHash digest2 = Hashcash.Hash(Stamp(answer2));
+#pragma warning restore CS0612
+            Assert.True(digest2.Satisfies(difficulty));
         }
 
         [Fact]