+++ Libplanet/Hashcash.cs
@@ -1,5 +1,6 @@
 #nullable enable
 using System;
+using System.Collections.Immutable;
 using System.Security.Cryptography;
 using System.Threading;
 using Libplanet.Blocks;
@@ -23,15 +24,62 @@ namespace Libplanet
         /// >proof-of-work system</a>, the total time an implementation elapses
         /// should not vary for different <paramref name="nonce"/>s.</para>
         /// </summary>
-        /// <param name="nonce">An arbitrary nonce for an attempt, provided
-        /// by <see cref="Hashcash.Answer(Stamp, long, CancellationToken)"/> method.</param>
+        /// <param name="nonce">An arbitrary nonce for an attempt, provided by
+        /// <see cref="Hashcash.Answer(Stamp, HashAlgorithmType, long, CancellationToken)"/> method.
+        /// </param>
         /// <returns>A <see cref="byte"/> array determined from the given
         /// <paramref name="nonce"/>.  It should return consistently
         /// an equivalent array for equivalent <paramref name="nonce"/>
         /// values.</returns>
-        /// <seealso cref="Hashcash.Answer(Stamp, long, CancellationToken)"/>
+        /// <seealso cref="Hashcash.Answer(Stamp, HashAlgorithmType, long, CancellationToken)"/>
         /// <seealso cref="Nonce"/>
         public delegate byte[] Stamp(Nonce nonce);
+
+        /// <summary>
+        /// Finds a <see cref="Nonce"/> that satisfies the given
+        /// <paramref name="difficulty"/>.  This process is so-called
+        /// &#x0201c;<a
+        /// href="https://en.wikipedia.org/wiki/Cryptocurrency#Mining"
+        /// >mining</a>&#x0201d;.
+        /// </summary>
+        /// <param name="stamp">A callback to get a &#x0201c;stamp&#x0201d;
+        /// which is a <see cref="byte"/> array determined from a given
+        /// <see cref="Nonce"/> value.</param>
+        /// <param name="hashAlgorithmType">The hash algorithm to use.</param>
+        /// <param name="difficulty">A number to calculate the target number
+        /// for which the returned answer should be less than.</param>
+        /// <param name="cancellationToken">
+        /// A cancellation token used to propagate notification that this
+        /// operation should be canceled.
+        /// </param>
+        /// <returns>A pair of <see cref="Nonce"/> value which satisfies the
+        /// given <paramref name="difficulty"/>, and the succeeded hash
+        /// digest.</returns>
+        /// <seealso cref="Stamp"/>
+        public static (Nonce Nonce, ImmutableArray<byte> Digest) Answer(
+            Stamp stamp,
+            HashAlgorithmType hashAlgorithmType,
+            long difficulty,
+            CancellationToken cancellationToken = default
+        )
+        {
+            var nonceBytes = new byte[10];
+            var random = new Random();
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                random.NextBytes(nonceBytes);
+                var nonce = new Nonce(nonceBytes);
+
+                // FIXME: .Satisfies() method should be moved to other class.
+                var digest = new BlockHash(hashAlgorithmType.Digest(stamp(nonce)));
+                if (digest.Satisfies(difficulty))
+                {
+                    return (nonce, digest.ByteArray);
+                }
+            }
+
+            throw new OperationCanceledException(cancellationToken);
+        }
 
         /// <summary>
         /// Finds a <see cref="Nonce"/> that satisfies the given
@@ -52,27 +100,13 @@ namespace Libplanet
         /// <returns>A <see cref="Nonce"/> value which satisfies the given
         /// <paramref name="difficulty"/>.</returns>
         /// <seealso cref="Stamp"/>
+        [Obsolete]
         public static Nonce Answer(
             Stamp stamp,
             long difficulty,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            var nonceBytes = new byte[10];
-            var random = new Random();
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                random.NextBytes(nonceBytes);
-                var nonce = new Nonce(nonceBytes);
-                var digest = Hash(stamp(nonce));
-
-                if (digest.Satisfies(difficulty))
-                {
-                    return nonce;
-                }
-            }
-
-            throw new OperationCanceledException(cancellationToken);
-        }
+            CancellationToken cancellationToken = default
+        ) =>
+            Answer(stamp, HashAlgorithmType.Of<SHA256>(), difficulty, cancellationToken).Nonce;
 
         /// <summary>
         /// Calculates a SHA-256 digest from the given <paramref name="bytes"/>.
@@ -81,6 +115,7 @@ namespace Libplanet
         /// its hash digest.</param>
         /// <returns>A deterministic digest of the given
         /// <paramref name="bytes"/>.</returns>
+        [Obsolete]
         public static BlockHash Hash(byte[] bytes) =>
             BlockHash.FromHashDigest(HashDigest<SHA256>.DeriveFrom(bytes));
     }