+++ CHANGES.md
@@ -195,6 +195,8 @@ To be released.
  -  Added `Currency(IValue)` overloaded constructor.  [[#1289]]
  -  Added `Currency.Serialize()` method.  [[#1289]]
  -  Added `ByteUtil.TimingSafelyCompare()` method.  [[#1314], [#1352]]
+ -  Added `Hashcash.Answer(Stamp, HashAlgorithm, long, CancellationToken)`
+    overloaded method.  [[#1314], [#1352]]
 
 ### Behavioral changes